requirements.txt
fastapi
uvicorn
joblib
pandas
numpy
xgboost
shap
anthropic
python-dotenv
streamlit
pyarrow
scikit-learn
requests
